Added in unlink check so it doesn't error.

+++ src/eops.coffee
@@ -58,7 +58,7 @@ do_hash = (program) ->
   return hashed_email
 
 program
-  .version('2.1.6')
+  .version('2.2.0')
   .usage('[options] <emails.csv>')
   .option('-c --compare <file>', 'hashed emails (already_hashed.csv)')
   .option('-e --case <alter>', 'whether to upper or lower case the email before hashing (upper, lower, as-is)', /^(upper|lower|as\-is)$/i, 'as-is')
@@ -163,7 +163,8 @@ else
     .on 'error', (err) ->
       console.log err
     .on 'end', () ->
-      if program.output
+      output_file_check = fs.statSync(program.output)
+      if program.output && output_file_check.isFile()
         fs.unlinkSync(program.output)
       fs.createReadStream(program.args[0])
         .pipe csv()
@@ -181,7 +182,8 @@ else
         .on 'error', (err) ->
           console.log err
   else
-    if program.output
+    output_file_check = fs.statSync(program.output)
+    if program.output && output_file_check.isFile()
       fs.unlinkSync(program.output)
       fs.appendFileSync(program.output, 'email'+"\r\n")
     fs.createReadStream program.args[0]
